Merges follower/following views
They were almost identical!
