+++ bookwyrm/templates/user/relationships/layout.html
@@ -6,11 +6,11 @@
 {% with user|username as username %}
 <nav class="tabs">
     <ul>
-        {% url 'user-followers' user|username as url %}
+        {% url 'user-relationships' user|username 'followers' as url %}
         <li{% if url == request.path or url == request.path|add:'/' %} class="is-active"{% endif %}>
             <a href="{{ url }}">{% trans "Followers" %}</a>
         </li>
-        {% url 'user-following' user|username as url %}
+        {% url 'user-relationships' user|username 'following' as url %}
         <li{% if url == request.path or url == request.path|add:'/' %} class="is-active"{% endif %}>
             <a href="{{ url }}">{% trans "Following" %}</a>
         </li>
