+++ bookwyrm/urls.py
@@ -376,14 +376,9 @@ urlpatterns = [
     re_path(rf"^@(?P<username>{regex.USERNAME})$", views.user_redirect),
     re_path(rf"{USER_PATH}/rss/?$", views.rss_feed.RssFeed(), name="user-rss"),
     re_path(
-        rf"{USER_PATH}/followers(.json)?/?$",
-        views.Followers.as_view(),
-        name="user-followers",
-    ),
-    re_path(
-        rf"{USER_PATH}/following(.json)?/?$",
-        views.Following.as_view(),
-        name="user-following",
+        rf"{USER_PATH}/(?P<direction>(followers|following))(.json)?/?$",
+        views.Relationships.as_view(),
+        name="user-relationships",
     ),
     re_path(r"^hide-suggestions/?$", views.hide_suggestions, name="hide-suggestions"),
     # groups
