+++ bookwyrm/views/__init__.py
@@ -127,7 +127,8 @@ from .setup import InstanceConfig, CreateAdmin
 from .status import CreateStatus, EditStatus, DeleteStatus, update_progress
 from .status import edit_readthrough
 from .updates import get_notification_count, get_unread_status_string
-from .user import User, Followers, Following, hide_suggestions, user_redirect
+from .user import User, hide_suggestions, user_redirect
+from .relationships import Relationships
 from .wellknown import *
 from .annual_summary import (
     AnnualSummary,
