+++ bookwyrm/views/user.py
@@ -1,8 +1,6 @@
-""" non-interactive pages """
+""" The user profile """
 from django.contrib.auth.decorators import login_required
-from django.core.exceptions import PermissionDenied
 from django.core.paginator import Paginator
-from django.db.models import Q, Count
 from django.http import Http404
 from django.shortcuts import redirect
 from django.template.response import TemplateResponse
@@ -100,62 +98,6 @@ class User(View):
         }
 
         return TemplateResponse(request, "user/user.html", data)
-
-
-class Followers(View):
-    """list of followers view"""
-
-    def get(self, request, username):
-        """list of followers"""
-        user = get_user_from_username(request.user, username)
-
-        if is_api_request(request):
-            return ActivitypubResponse(user.to_followers_activity(**request.GET))
-
-        if user.hide_follows and user != request.user:
-            raise PermissionDenied()
-
-        followers = annotate_if_follows(request.user, user.followers)
-        paginated = Paginator(followers.all(), PAGE_LENGTH)
-        data = {
-            "user": user,
-            "is_self": request.user.id == user.id,
-            "follow_list": paginated.get_page(request.GET.get("page")),
-        }
-        return TemplateResponse(request, "user/relationships/followers.html", data)
-
-
-class Following(View):
-    """list of following view"""
-
-    def get(self, request, username):
-        """list of followers"""
-        user = get_user_from_username(request.user, username)
-
-        if is_api_request(request):
-            return ActivitypubResponse(user.to_following_activity(**request.GET))
-
-        if user.hide_follows and user != request.user:
-            raise PermissionDenied()
-
-        following = annotate_if_follows(request.user, user.following)
-        paginated = Paginator(following.all(), PAGE_LENGTH)
-        data = {
-            "user": user,
-            "is_self": request.user.id == user.id,
-            "follow_list": paginated.get_page(request.GET.get("page")),
-        }
-        return TemplateResponse(request, "user/relationships/following.html", data)
-
-
-def annotate_if_follows(user, queryset):
-    """Sort a list of users by if you follow them"""
-    if not user.is_authenticated:
-        return queryset.order_by("-created_date")
-
-    return queryset.annotate(
-        request_user_follows=Count("followers", filter=Q(followers=user))
-    ).order_by("-request_user_follows", "-created_date")
 
 
 @require_POST
